Move flowgraph into flowc as internal graph module (#2740)
Move the flowgraph crate (~946 lines) into flowc/src/lib/graph/ as an
internal module behind a `graph` feature flag (enabled by default).

- flowgraph was only used by flowc for SVG graph output with -g flag
- No other crate depended on it
- Removes a workspace crate, simplifying publishing and maintenance
- SVG generation is now optional via the graph feature

Removed unused COLOR_BACKGROUND constant and suppressed dead_code
warning on NodeLayout.name field.
