+++ flowc/src/lib/lib.rs
@@ -22,3 +22,7 @@ pub mod info;
 /// contains `errors::Error` that other modules in this crate will `use errors::*;`
 /// to get access to everything `error_chain` creates.
 pub mod errors;
+
+/// Graph layout and SVG rendering for flow diagrams
+#[cfg(feature = "graph")]
+pub mod graph;
