--- flowgraph/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "flowgraph"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-description = "Graph layout and SVG rendering for flow programs"
-
-[lints]
-workspace = true
-
-[dependencies]
-flowcore = { path = "../flowcore", version = "1.0.0" }
-svg = "0.18"
-log = "0.4"
-url = "2.2"
